pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Project 'Differences report"
authors = ["Daniil Bagaturiya <daniilbagaturiya@ya.ru>"]
readme = "README.md"
packages = [
  { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0.1"

[tool.poetry.scripts]
gendiff = 'gendiff.scripts.gen:main'

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
flake8 = "^6.1.0"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
